label shape

+++ train.py
@@ -162,7 +162,7 @@ def main(args):
             logits = model(cluster, cluster.ndata['feat'])
             train_y = cluster.ndata['label']
             if multitask == False:
-                train_y = F.one_hot(train_y, num_classes=n_classes)
+                train_y = F.one_hot(train_y, num_classes=n_classes).float()
             print(logits.shape, train_y.shape)
             print(train_y)
             loss = criterion(logits, train_y)    # 计算损失值
